should all work need to change to int in token

+++ tokens.hpp
@@ -42,6 +42,6 @@
   };
   extern int yylineno;
   extern char* yytext;
-  extern int yyleng;
+  extern unsigned long yyleng;
   extern int yylex();
 #endif /* TOKENS_HPP_ */
